Move note usecase to each folder
