+++ app/src/main/java/com/laam/laamnotes/framework/interactors/NoteDetailInteractors.kt
@@ -1,9 +1,8 @@
 package com.laam.laamnotes.framework.interactors
 
-import com.laam.core.usecase.AddNote
-import com.laam.core.usecase.GetAllNotes
-import com.laam.core.usecase.GetNote
-import com.laam.core.usecase.RemoveNote
+import com.laam.core.usecase.note.AddNote
+import com.laam.core.usecase.note.GetNote
+import com.laam.core.usecase.note.RemoveNote
 
 data class NoteDetailInteractors(
     val addNote: AddNote,
