+++ app/src/main/java/com/laam/laamnotes/framework/interactors/NoteListInteractors.kt
@@ -1,6 +1,6 @@
 package com.laam.laamnotes.framework.interactors
 
-import com.laam.core.usecase.GetAllNotes
+import com.laam.core.usecase.note.GetAllNotes
 
 data class NoteListInteractors(
     val getAllNotes: GetAllNotes
