+++ core/src/main/java/com/laam/core/usecase/note/AddNote.kt
@@ -1,7 +1,7 @@
-package com.laam.core.usecase
+package com.laam.core.usecase.note
 
 import com.laam.core.data.Note
-import com.laam.core.repository.NoteRepository
+import com.laam.core.repository.note.NoteRepository
 
 class AddNote(private val repository: NoteRepository) {
