+++ core/src/main/java/com/laam/core/usecase/note/GetNote.kt
@@ -1,6 +1,6 @@
-package com.laam.core.usecase
+package com.laam.core.usecase.note
 
-import com.laam.core.repository.NoteRepository
+import com.laam.core.repository.note.NoteRepository
 
 class GetNote(private val repository: NoteRepository) {
